+++ csrc/bindings.h
@@ -35,8 +35,8 @@ project_gaussians_forward_tensor(
     torch::Tensor projmat,
     const float fx,
     const float fy,
-    const float img_height,
-    const float img_width,
+    const unsigned img_height,
+    const unsigned img_width,
     const std::tuple<int, int, int> tile_bounds
 );
 
@@ -56,8 +56,8 @@ project_gaussians_backward_tensor(
     torch::Tensor projmat,
     const float fx,
     const float fy,
-    const float img_height,
-    const float img_width,
+    const unsigned img_height,
+    const unsigned img_width,
     torch::Tensor cov3d,
     torch::Tensor radii,
     torch::Tensor conics,
